LPS-53630 Move upgrade process to the module
